+++ lib/src/lib.rs
@@ -90,6 +90,7 @@ pub mod repo;
 pub mod repo_path;
 pub mod revset;
 mod revset_parser;
+pub mod revset_util;
 pub mod rewrite;
 #[cfg(feature = "testing")]
 pub mod secret_backend;
